Increase library loading speed.
One of the most expensive operations is hashing the entire file to
determine the track ID, so we allow many different types of UUIDs
(which all *should* uniquely identify the track) to be a track ID.
There is a slight risk of collision since three different ID types
may be in use, but this is probably fine.

The set of track attributes that is stored is also restricted.